Move MongoPhotoGridDataSource to separate file and add pagination support
- Move MongoPhotoGridDataSource out of reusable PhotoGrid files as requested
- Remove MongoDB imports from reusable components
- Add pagination support to loadItems function with offset and length parameters
- Update MongoPhotoGridDataSource to support appending items for pagination
- Maintain backward compatibility with default parameters
